:hammer: Move Modal component to global file and update import
